+++ src/segmentation/segmentation.py
@@ -75,7 +75,22 @@ def convert_to_labels2d(slice):
     return new_image
 
     # return drawn_contours
+def automatic_contours(ground_truth):
+    result = np.zeros(ground_truth.shape,dtype=bool)
+    target = int(len(ground_truth)/2)
+    result[target] = create_contour(ground_truth[target])
+    return result
 
+def create_contour(ground_truth):
+    if np.count_nonzero(ground_truth) == 0:
+        return ground_truth
+    result = np.zeros(ground_truth.shape,dtype=bool)
+    im = ground_truth.astype(np.uint8)
+    contours, hierarchy = cv.findContours(im, cv.RETR_TREE, cv.CHAIN_APPROX_NONE)
+    for contour in contours:
+        for point in contour:
+            result[point[0][1], point[0][0]] = True
+    return result
 
 def segment(image, seed_points):
     print("Generating segment...")
